remove replying comments
