+++ Gemfile
@@ -15,9 +15,6 @@ gem "mission_control-jobs"
 gem "activerecord-typedstore"
 gem "importmap-rails", "~> 2.0"
 gem "propshaft"
-gem "scenic"
-gem "scenic-mysql_adapter"
-gem "scenic_sqlite_adapter"
 gem "sentry-rails"
 gem "typeid"
