+++ app/models/user.rb
@@ -604,28 +604,8 @@ class User < ApplicationRecord
     true
   end
 
-  def unread_message_count
-    @unread_message_count ||= Keystore.value_for("user:#{id}:unread_messages").to_i
-  end
-
-  def update_unread_message_count!
-    @unread_message_count = received_messages.unread.count
-    Keystore.put("user:#{id}:unread_messages", @unread_message_count)
-  end
-
-  def clear_unread_replies!
-    Rails.cache.delete("user:#{id}:unread_replies")
-  end
-
-  def unread_replies_count
-    @unread_replies_count ||=
-      Rails.cache.fetch("user:#{id}:unread_replies", expires_in: 2.minutes) {
-        ReplyingComment.where(user_id: id, is_unread: true).count
-      }
-  end
-
   def inbox_count
-    unread_message_count + unread_replies_count
+    notifications.where(read_at: nil).count
   end
 
   def votes_for_others
