+++ app/views/messages/_subnav.html.erb
@@ -1,12 +1,7 @@
 <%# locals: () -%>
 <% content_for :subnav do %>
-  <%= link_to_different_page('Unread',   replies_unread_path) %>
-  <%= link_to_different_page('All',      replies_path) %>
-  <%= link_to_different_page('Comments', replies_comments_path) %>
-  <%= link_to_different_page('Stories',  replies_stories_path) %>
-  |
   <%= link_to_different_page 'Messages', messages_path %>
   <%= link_to_different_page 'Sent', messages_sent_path %>
-  <%= link_to_different_page 'All Notifications', notifications_path %>
-  <%= link_to_different_page 'Unread Notifications', notifications_unread_path %>
+  <%= link_to_different_page 'All Notifications', inbox_all_path %>
+  <%= link_to_different_page 'Unread Notifications', inbox_unread_path %>
 <% end %>
